functional_tests: move RPC API to utils, it is not test specific
