+++ tests/functional_tests/cold_signing.py
@@ -33,8 +33,8 @@ import time
 """Test cold tx signing
 """
 
-from test_framework.daemon import Daemon
-from test_framework.wallet import Wallet
+from framework.daemon import Daemon
+from framework.wallet import Wallet
 
 class ColdSigningTest():
     def run_test(self):
