+++ tests/functional_tests/multisig.py
@@ -33,8 +33,8 @@ import time
 """Test multisig transfers
 """
 
-from test_framework.daemon import Daemon
-from test_framework.wallet import Wallet
+from framework.daemon import Daemon
+from framework.wallet import Wallet
 
 class MultisigTest():
     def run_test(self):
