+++ tests/functional_tests/sign_message.py
@@ -38,7 +38,7 @@ Test the following RPCs:
 
 """
 
-from test_framework.wallet import Wallet
+from framework.wallet import Wallet
 
 class MessageSigningTest():
     def run_test(self):
